feat: Added AMR goal setting functionality
A new feature has been added to set the Autonomous Mobile Robot (AMR) goal. This includes a new action class, `AMRGoalAction`, which allows users to define an AMR goal on the map. The action is now available in the tools menu. Additionally, a corresponding icon image for this action has been included.