feat: Added MovableAreaAction functionality
A new feature, MovableAreaAction, has been added. This includes a new icon and the corresponding Java class. The feature allows users to set a movable area on the map. It also checks if the selected OSM collection is editable and contains only one item which is an area. The ToolsPlugin class has been updated to include this action in its menu.