feat: Added safety area functionality
A new feature has been added to set a safety area. This includes the creation of a new action class, SafetyAreaAction, which allows users to define specific properties for the safety area. An icon for this action has also been included. The new action is now part of the main menu in the ToolsPlugin class.