feat: Added NonStopAreaAction functionality
A new feature has been added to the toolset. This includes a new action, NonStopAreaAction, which allows users to set non-stop areas. The implementation involves creating a new Java class and adding an icon for this action. The feature also includes changes in properties of selected areas in the map. Furthermore, the NonStopAreaAction has been integrated into the main menu of the application.