feat: Added SFTP functionality to plugin
In this update, we've introduced SFTP support to the plugin. This includes adding a new class path in the build file and creating new Java classes for handling SFTP actions and dialogs. We've also added an icon for the SFTP action and made necessary changes in the main plugin file to include this new feature in the menu.